+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Added support to get image size for SVG vector images [#3533](https://github.com/getgrav/grav/pull/3533)
     * Added XSS check for uploaded SVG files before they get stored
     * Fixed phpstan issues (All level 2, Framework level 5)
+    * Added a `system.legacy_url_root_behavior` config option to enable legacy/broken `Utils::url()` behavior
 2. [](#bugfix)
     * Fixed `'mbstring' extension is not loaded` error, use Polyfill instead [#3504](https://github.com/getgrav/grav/pull/3504)
     * Fixed new `Utils::pathinfo()` and `Utils::basename()` being too strict for legacy use [#3542](https://github.com/getgrav/grav/issues/3542)